Move DocumentHead to correct folder with /documents
